last

+++ InterMove/src/main/java/com/example/intermove/Controllers/EventsAndComplaints/ClaimController.java
@@ -29,7 +29,7 @@ public class ClaimController {
 
     public Claim addReclamation(@RequestParam("objet") String objet, @RequestParam ("message") String message,
                                 @RequestParam ("typeClaim") TypeClaim typeClaim
-            , @RequestParam("upload") MultipartFile upload ,@PathVariable Long id) throws IOException, URISyntaxException
+            , @RequestParam("upload") MultipartFile upload ,@PathVariable int id) throws IOException, URISyntaxException
     {
 
         Claim claim = new Claim() ;
@@ -101,7 +101,7 @@ public class ClaimController {
         return new ModelAndView("index");
     }
     @PostMapping("/{id}/send-email")
-    public ResponseEntity<String> sendEmail(@PathVariable Long id) {
+    public ResponseEntity<String> sendEmail(@PathVariable int id) {
         reclamationService.sendEmail(id);
         return ResponseEntity.ok("E-mail envoyé avec succès.");
     }
